chore: Move post listing endpoint to endpoints section of app
